+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- (#14) `--dump-db <path>` dumps to `<path>` instead to the `dumped.txt`
 
 ## [0.3.0] - 2020-11-26
 ### Added